backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-dotenv==1.0.1
langchain==0.3.15
langchain-google-genai==2.0.5
langgraph==0.2.62
pydantic==2.10.4
playwright==1.48.0

